# LoginRegistration_MiniProject/backend/models/User.js
// models/User.js
const mongoose = require("mongoose");

const UserSchema = new mongoose.Schema({
  username: {
    type: String,
    required: true,
    unique: true,
  },
  email: {
    type: String,
    required: true,
    unique: true,
  },
  password: {
    type: String,
    required: true,
  },
  profile: {
    firstName: String,
    lastName: String,
    phoneNumber: String,
  },
});

module.exports = mongoose.model("User", UserSchema);